requirements.txt
asgiref==3.2.7
astroid==2.4.2
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==2.2.5
django-appconf==1.0.4
django-crispy-forms==1.9.1
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber_field
django-select2==7.4.2
graphviz==0.13.2
html5lib==1.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.5
Pillow==7.2.0
pylint==2.5.3
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.49
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
virtualenv==16.7.9
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
django-csvimport==2.16
paywix
